+++ src_tests/test_minimize_IPNewton.cc
@@ -329,7 +329,8 @@ void print_summary_table()
     accumulated_gradient_steps += r.gradient_steps;
   }
 
-  auto perc = 100.0 * converged_tests / std::max<size_t>( total_tests, 1 );
+  auto perc =
+    100.0 * static_cast<double>( converged_tests ) / static_cast<double>( std::max<size_t>( total_tests, 1 ) );
 
   fmt::print( fmt::fg( fmt::color::light_blue ), "\n📊 Statistics:\n" );
   fmt::print( "   • Total problems: {}\n", total_tests );
